requirements.txt
numpy==1.26.1
Pillow==10.0.1
